requirements.txt
setuptools==45.2.0
sqlparse==0.4.1
asgiref==3.4.1
pytz==2019.3
django==3.2.5
discord==1.7.3
aiohttp==3.7.4.post0
python-dotenv==0.18.0
